# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VentasFix API",
        "description": "API REST del backoffice de VentasFix para la gestión de usuarios, productos y clientes. Pensada para el consumo desde aplicaciones de terceros (por ejemplo, el sistema de gestión Softland). Todos los endpoints protegidos requieren un token Bearer obtenido en /api/login.",
        "contact": {
            "name": "Felipe Jiménez",
            "email": "felipeandresjb99@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar todos los clientes",
                "operationId": "3ab1bd0bc1c6d13508917ec1f71f52f5",
                "responses": {
                    "200": {
                        "description": "Listado de clientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Cliente"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear un nuevo cliente",
                "operationId": "22be7ccdd26efb899a619bd891090f4f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rut_empresa",
                                    "rubro",
                                    "razon_social",
                                    "telefono",
                                    "direccion",
                                    "nombre_contacto",
                                    "email_contacto"
                                ],
                                "properties": {
                                    "rut_empresa": {
                                        "type": "string",
                                        "example": "76.123.456-7"
                                    },
                                    "rubro": {
                                        "type": "string",
                                        "example": "Retail"
                                    },
                                    "razon_social": {
                                        "type": "string",
                                        "example": "Comercial Los Andes SpA"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "+56 9 1234 5678"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Siempre Viva 123"
                                    },
                                    "nombre_contacto": {
                                        "type": "string",
                                        "example": "María Pérez"
                                    },
                                    "email_contacto": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria@losandes.cl"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente creado correctamente."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/clientes/{cliente}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener un cliente por su ID",
                "operationId": "c0fd13ba043d2ca99f23be8586cdd267",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar un cliente por su ID",
                "operationId": "02e774a78bba9aee8339fcb5bba74ea9",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rut_empresa",
                                    "rubro",
                                    "razon_social",
                                    "telefono",
                                    "direccion",
                                    "nombre_contacto",
                                    "email_contacto"
                                ],
                                "properties": {
                                    "rut_empresa": {
                                        "type": "string",
                                        "example": "76.123.456-7"
                                    },
                                    "rubro": {
                                        "type": "string",
                                        "example": "Retail"
                                    },
                                    "razon_social": {
                                        "type": "string",
                                        "example": "Comercial Los Andes SpA"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "+56 9 1234 5678"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Siempre Viva 123"
                                    },
                                    "nombre_contacto": {
                                        "type": "string",
                                        "example": "María Pérez"
                                    },
                                    "email_contacto": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria@losandes.cl"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente actualizado correctamente."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Eliminar un cliente por su ID",
                "operationId": "34091ea4b00349dde0aad1b503809bdb",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cliente eliminado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Cliente": {
                "title": "Cliente",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "rut_empresa": {
                        "type": "string",
                        "example": "76.123.456-7"
                    },
                    "rubro": {
                        "type": "string",
                        "example": "Retail"
                    },
                    "razon_social": {
                        "type": "string",
                        "example": "Comercial Los Andes SpA"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "+56 9 1234 5678"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Av. Siempre Viva 123"
                    },
                    "nombre_contacto": {
                        "type": "string",
                        "example": "María Pérez"
                    },
                    "email_contacto": {
                        "type": "string",
                        "format": "email",
                        "example": "maria@losandes.cl"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "title": "Error de validación",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "El campo email es obligatorio."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Introduce el token entregado por /api/login como: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}